Crate ratatui_grid_canvas_widget

This crate provides widgets for rendering colored grids in Ratatui.

Feature Flags (all on by default)

binary

You might want this disabled if you’re not using it since it brings in dependencies (bitvec and whatever it depends on)

color

Honestly, this is just a feature because binary is. This doesn’t bring in any dependencies. But if you’re using binary grids and not color grids and you want to minimize your compile time…

alloc

This enables grid::alloc which contains resizeable heap-allocated grids.

Color Grids

Instead of storing [Color]s, it takes in a generic type that implements ToColor so you can use your own type for the grid and use it to implement logic better, and then just add the trait so it knows how to turn it into a [Color]. If you don’t provide a type, it will use [Color] by default.

Binary Grids

At the cost of being limited to 2 colors instead of 16777216, these have 2 advantages over color grids:

Smaller memory usage, since each cell only takes up 1 bit

Higher resolution options using quadrants, sextants, or braille.

Coordinates

Like Ratatui, the coordinate system in the grids runs left to right, top to bottom, with the origin (0, 0) in the top left corner.

Modules

grid

widget

Traits

BinaryGrid

ColorGrid

ToColor