ratatui_grid_canvas_widget::widget

Module color

Structs

DoubleFullBlockColorGridWidget

Uses 2 Unicode full blocks which each fill an entire terminal cell. They are placed next to each other on the same row, not the same column.

FullBlockColorGridWidget

Uses Unicode full blocks which fill an entire terminal cell.

HalfBlockColorGridWidget

Uses Unicode half blocks and full blocks which allows you to have 2 colors per terminal cell, doubling the vertical resolution.

HorizontalHalfBlockColorGridWidget

Uses Unicode half blocks and full blocks which allows you to have 2 colors per terminal cell, doubling the horizontal resolution, albeit with perhaps weirdly skinny “pixels”.